improve pm module
