+++ src/modules/teams/validation/schema.ts
@@ -6,12 +6,12 @@ export const createTeamSchema = z.object({
   name: z.string().min(1, "نام تیم الزامی است."),
   description: z.string().optional(),
   members: z.array(z.object({ id: z.number() })).optional(),
-  parentId: z.number().optional().nullable(),
+ parent: z.object({ id: z.number() }).optional().nullable(),
 });
 
 export const updateTeamSchema = z.object({
   name: z.string().min(1, "نام تیم الزامی است.").optional(),
   description: z.string().optional(),
   members: z.array(z.object({ id: z.number() })).optional(),
-  parentId: z.number().optional().nullable(),
+   parent: z.object({ id: z.number() }).optional().nullable(),
 });
