+++ src/modules/teams/views/page.tsx
@@ -2,7 +2,8 @@
 "use client";
 
 import Loading from "@/@Client/Components/common/Loading";
-import DataTableWrapper3 from "@/@Client/Components/wrappers/DataTableWrapper3";
+
+import DataTableWrapper from "@/@Client/Components/wrappers/DataTableWrapper2";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import TeamTree from "../components/TeamTree";
 import { columns } from "../data/table";
@@ -89,7 +90,7 @@ export default function IndexPage() {
 
           {/* ستون اصلی (جدول تیم‌ها) */}
           <div className="lg:col-span-3">
-            <DataTableWrapper3<TeamWithRelations>
+            <DataTableWrapper<TeamWithRelations>
               columns={columns}
               createUrl="/dashboard/teams/create"
               loading={loading}
